Moving files into src directory, initial setup for packaging
