+++ .gitignore
@@ -1,2 +1,6 @@
 **/__pycache__/
 venv/
+
+# Packaging folders
+dist/
+build/
