+++ src/metaeditor_safetensors/main.py
@@ -1,5 +1,5 @@
 from editor import SafetensorsEditor
 
 if __name__ == "__main__":
-    app = SafetensorsEditor()
-    app.mainloop()
+	app = SafetensorsEditor()
+	app.mainloop()
